Refactor font
