+++ src/font/font.h
@@ -15,7 +15,7 @@
     if (!(*(int *)PyFONT_C_API[2])) \
     return RAISE(pgExc_SDLError, "font system not initialized")
 
-#include "include/pygame_font.h"
+#include "pygame_font.h"
 
 #define PYGAMEAPI_FONT_NUMSLOTS 3
